move avatar input and chain select
